+++ lib/numerus.dart
@@ -1,9 +1,3 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
 library numerus;
 
-export 'src/int_extension.dart';
-export 'src/string_extension.dart';
-
-// TODO: Export any libraries intended for clients of this package.
+export 'roman/roman.dart';