+++ lib/roman/roman.dart
@@ -1,0 +1,4 @@
+export 'roman_config.dart';
+export 'roman_numerals.dart';
+export '../src/roman/roman_int_extension.dart';
+export '../src/roman/roman_string_extension.dart';