+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 2.0.0
+
+ - Support for Vinculum or Apostrophus (both compact with Unicode
+   and tally marks)
+- **NOTE**: this change means the maximum value of common/default config is
+     3,999; 399,999 for 'compact' apostrophus', and 3,999,999 for both
+     tally mark apostrophus and vinculum styles.
+ - Housekeeping for bigger Dart update, as well as `test` and `analyze`
+   cleanup.
+ - Improved code coverage report generation to be "pure Dart" and scripted.
+
 ## 1.1.1
 
  - Full null-safety support.